fix 537
the function _is_ngsi_array was not computing correctly the data type casting in case the attribute type was not "Array" or "StructuredValue"

also, the call was returning 200, despite the error given that "raise" was not issued. this last thing should be checked in better detail with ©c0c0n3 since it may have been on purpose.